bridge/pyproject.toml
[project]
name = "ra2-bridge"
version = "0.1.0"
description = "Lutron RadioRA2 to MQTT bridge for Homey"
requires-python = ">=3.12"
dependencies = [
    "telnetlib3>=2.0,<3.0",
    "aiomqtt>=2.0,<3.0",
    "aiohttp>=3.9,<4.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
